fix: 카테고리 고정되어 있는 거 수정
