+++ frontend/src/domains/user/OnBoarding/OnBoarding.tsx
@@ -14,12 +14,13 @@ import {
 } from '@/domains/user/OnBoarding/OnBoarding.styles';
 import { useAppTheme } from '@/hooks/useAppTheme';
 import { Analytics, onboardingEvents } from '@/analytics';
-import { CategoryType } from '@/analytics/types';
 import { useOrganizationId } from '@/domains/hooks/useOrganizationId';
 import useNavigation from '@/domains/hooks/useNavigation';
+import { createCategoryIconPairs } from '@/domains/utils/createCategoryList';
+import { CategoryListType } from '@/constants/categoryList';
 
 interface OnBoardingProps {
-  onCategoryClick: (newCategory: CategoryType) => void;
+  onCategoryClick: (newCategory: CategoryListType) => void;
 }
 
 export default function OnBoarding({ onCategoryClick }: OnBoardingProps) {
@@ -27,9 +28,11 @@ export default function OnBoarding({ onCategoryClick }: OnBoardingProps) {
   const { goPath } = useNavigation();
   const { organizationId } = useOrganizationId();
 
-  const { groupName } = useOrganizationName({
+  const { groupName, categories } = useOrganizationName({
     organizationId,
   });
+
+  const categoryIconPairs = createCategoryIconPairs(categories);
 
   const handleViewSuggestionsClick = () => {
     Analytics.track(onboardingEvents.viewSuggestionsFromOnboarding());
@@ -49,26 +52,14 @@ export default function OnBoarding({ onCategoryClick }: OnBoardingProps) {
           <p css={question(theme)}>건의하고 싶은 카테고리를 선택해주세요</p>
         </div>
         <div css={buttonContainer}>
-          <CategoryButton
-            icon='🚨'
-            text='신고'
-            onClick={() => onCategoryClick('신고')}
-          />
-          <CategoryButton
-            icon='🙋‍♀️'
-            text='질문'
-            onClick={() => onCategoryClick('질문')}
-          />
-          <CategoryButton
-            icon='💬'
-            text='건의'
-            onClick={() => onCategoryClick('건의')}
-          />
-          <CategoryButton
-            icon='💡'
-            text='기타'
-            onClick={() => onCategoryClick('기타')}
-          />
+          {categoryIconPairs.map((category) => (
+            <CategoryButton
+              key={category.category}
+              icon={category.icon}
+              text={category.category}
+              onClick={() => onCategoryClick(category.category)}
+            />
+          ))}
         </div>
       </div>
       <BasicButton
